@@ -216,11 +216,7 @@ where
                 let (class_element_name, method) =
                     AsyncMethod::new(self.allow_yield, self.allow_await).parse(cursor, interner)?;
 
-                let property_name = if let property::ClassElementName::PropertyName(property_name) =
-                    class_element_name
-                {
-                    property_name
-                } else {
+                let property::ClassElementName::PropertyName(property_name) = class_element_name else {
                     return Err(ParseError::general(
                         "private identifiers not allowed in object literal",
                         position,
