@@ -261,20 +261,15 @@ where
                 let bindings = ObjectBindingPattern::new(self.allow_yield, self.allow_await)
                     .parse(cursor, interner)?;
 
-                let init = if let Some(t) = cursor.peek(0, interner)? {
-                    if *t.kind() == TokenKind::Punctuator(Punctuator::Assign) {
-                        Some(
-                            Initializer::new(
-                                None,
-                                self.allow_in,
-                                self.allow_yield,
-                                self.allow_await,
-                            )
+                let init = if cursor
+                    .peek(0, interner)?
+                    .filter(|t| *t.kind() == TokenKind::Punctuator(Punctuator::Assign))
+                    .is_some()
+                {
+                    Some(
+                        Initializer::new(None, self.allow_in, self.allow_yield, self.allow_await)
                             .parse(cursor, interner)?,
-                        )
-                    } else {
-                        None
-                    }
+                    )
                 } else {
                     None
                 };
@@ -294,20 +289,15 @@ where
                 let bindings = ArrayBindingPattern::new(self.allow_yield, self.allow_await)
                     .parse(cursor, interner)?;
 
-                let init = if let Some(t) = cursor.peek(0, interner)? {
-                    if *t.kind() == TokenKind::Punctuator(Punctuator::Assign) {
-                        Some(
-                            Initializer::new(
-                                None,
-                                self.allow_in,
-                                self.allow_yield,
-                                self.allow_await,
-                            )
+                let init = if cursor
+                    .peek(0, interner)?
+                    .filter(|t| *t.kind() == TokenKind::Punctuator(Punctuator::Assign))
+                    .is_some()
+                {
+                    Some(
+                        Initializer::new(None, self.allow_in, self.allow_yield, self.allow_await)
                             .parse(cursor, interner)?,
-                        )
-                    } else {
-                        None
-                    }
+                    )
                 } else {
                     None
                 };
@@ -334,20 +324,20 @@ where
                     )));
                 }
 
-                let init = if let Some(t) = cursor.peek(0, interner)? {
-                    if *t.kind() == TokenKind::Punctuator(Punctuator::Assign) {
-                        Some(
-                            Initializer::new(
-                                Some(ident),
-                                self.allow_in,
-                                self.allow_yield,
-                                self.allow_await,
-                            )
-                            .parse(cursor, interner)?,
+                let init = if cursor
+                    .peek(0, interner)?
+                    .filter(|t| *t.kind() == TokenKind::Punctuator(Punctuator::Assign))
+                    .is_some()
+                {
+                    Some(
+                        Initializer::new(
+                            Some(ident),
+                            self.allow_in,
+                            self.allow_yield,
+                            self.allow_await,
                         )
-                    } else {
-                        None
-                    }
+                        .parse(cursor, interner)?,
+                    )
                 } else {
                     None
                 };
